fixed delete menu

+++ FoodDeliveryApp/src/main/java/com/hexaware/hotpot/controller/RestaurantController.java
@@ -73,7 +73,7 @@ public class RestaurantController {
 
 	@DeleteMapping("/deleteMenu/{menuId}")
     @PreAuthorize("hasAuthority('admin') or hasAuthority('manager')")
-	public String deleteMenu(@PathVariable int menuId) {
+	public String deleteMenu(@PathVariable long menuId) {
 		service.deleteMenu(menuId);
 		return "menu deleted successfully";
 	}
